Apply forbidden term usage to all translation stages
