+++ .tooling/markdownlint/custom-formatting.th.js
@@ -1,61 +1,17 @@
-import { forEachLine, getLineMetadata } from "markdownlint-rule-helpers";
-import fs from "fs";
-import path from "path";
-const forbiddenTerms = fs
-  .readFileSync(
-    path.resolve(".tooling/markdownlint", "./forbidden_terms/thai.txt"),
-    "utf-8"
-  )
-  .split("\n")
-  .filter((l) => l && l.charAt(0) !== "#");
+import { applyCustomFormattingRules, readForbiddenTerms } from "./helpers.js";
 
 const customFormattingRules = [
   {
-    name: "missing space after closing curly quotes",
-    regexp: /[’”][^\s<’”—\.\[]/,
-  },
-  {
-    name: "no trailing space after 'ๆ' mai yamok",
-    regexp: /ๆ[^\s’”\[]/,
-  },
-  {
     name: "forbidden term",
-    regexp: new RegExp(`${forbiddenTerms.join("|")}`),
+    regexp: new RegExp(`${readForbiddenTerms("thai.txt").join("|")}`),
   },
 ];
-
-const clamp = (number, min, max) => {
-  return Math.max(min, Math.min(number, max));
-};
-
-const extractContext = (line, column) => {
-  const contextPadding = 10;
-  return line.substr(
-    clamp(column - contextPadding, 0, line.length - 1),
-    contextPadding * 2
-  );
-};
 
 const CustomFormatting = {
   names: ["custom-formatting-th"],
   description: "Custom formatting (th) rules",
   tags: ["style"],
-  function: (params, onError) => {
-    forEachLine(getLineMetadata(params), (line, lineIndex) => {
-      for (const rule of customFormattingRules) {
-        const { name, regexp } = rule;
-        const column = line.search(regexp);
-        if (column >= 0) {
-          onError({
-            lineNumber: lineIndex + 1,
-            detail: `Found ${name} at column ${column}`,
-            context: extractContext(line, column),
-            range: [column + 1, 1],
-          });
-        }
-      }
-    });
-  },
+  function: applyCustomFormattingRules(customFormattingRules),
 };
 
 export default CustomFormatting;
